fix a NPE in bufferactivity when we rotate quickly

+++ src/com/lekebilen/quasseldroid/gui/BufferActivity.java
@@ -143,7 +143,11 @@ public class BufferActivity extends ListActivity {
 
 		public void setBuffers(BufferCollection buffers){
 			this.bufferCollection = buffers;
-			bufferCollection.addObserver(this);
+			
+			if (buffers == null)
+				return;
+			
+			this.bufferCollection.addObserver(this);
 			notifyDataSetChanged();
 		}
 
@@ -223,6 +227,7 @@ public class BufferActivity extends ListActivity {
 		}
 
 		public void stopObserving() {
+			if (bufferCollection == null) return;
 			bufferCollection.deleteObserver(this);
 			
 		}
@@ -410,7 +415,8 @@ public class BufferActivity extends ListActivity {
 		if (isBound) {
 			Log.i(TAG, "Unbinding service");
 			bufferListAdapter.stopObserving();
-			boundConnService.unregisterStatusReceiver(statusReciver);
+			if (boundConnService != null)
+				boundConnService.unregisterStatusReceiver(statusReciver);
 			// Detach our existing connection.
 			unbindService(mConnection);
 			isBound = false;
